backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
torch==2.3.0
torchvision==0.18.0
transformers==4.43.0
timm==0.9.16
einops==0.8.0
Pillow==10.3.0
groq==0.9.0
httpx<0.28.0
